added test for file parsing, corrected program freeing
