+++ src/program/program.c
@@ -6,7 +6,7 @@
 /*   By: pandalaf <pandalaf@student.42wolfsburg.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/12 17:49:31 by pandalaf          #+#    #+#             */
-/*   Updated: 2023/01/20 05:07:20 by pandalaf         ###   ########.fr       */
+/*   Updated: 2023/01/20 13:59:38 by pandalaf         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
